Moved CaseUtilTest & resolved problems from moving CaseUtils
