+++ inspectit-ocelot-config/src/test/java/rocks/inspectit/ocelot/config/utils/CaseUtilTest.java
@@ -1,4 +1,4 @@
-package rocks.inspectit.ocelot.core.config.util;
+package rocks.inspectit.ocelot.config.utils;
 
 import org.junit.jupiter.api.Test;
